# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "ShapeShift",
      "description": "OpenAPI Schema for ShapeShift",
      "version": "0.0.1"
    },
    "paths": {
...
      "/api/v1/workout/{workout_id}": {
        "get": {
          "tags": ["workouts"],
          "summary": "Get a single workout",
          "operationId": "getWorkout",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/WorkoutOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["workouts"],
          "summary": "Update a workout",
          "operationId": "updateWorkout",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/WorkoutOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["workouts"],
          "summary": "Delete a workout",
          "operationId": "deleteWorkout",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/goal/": {
        "get": {
          "tags": ["goals"],
          "summary": "Get all goals",
          "operationId": "getGoals",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/GoalOut"
                    },
                    "type": "array",
                    "title": "Response Getgoals"
                  }
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        },
        "post": {
          "tags": ["goals"],
          "summary": "Create a goal",
          "operationId": "createGoal",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Goal"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      },
      "/api/v1/goal/{goal_id}": {
        "get": {
          "tags": ["goals"],
          "summary": "Get a single goal",
          "operationId": "getGoal",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "goal_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Goal Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GoalOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["goals"],
          "summary": "Update a goal",
          "operationId": "updateGoals",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "goal_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Goal Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GoalOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["goals"],
          "summary": "Delete a goal",
          "operationId": "deleteGoal",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "goal_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Goal Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/session/": {
        "get": {
          "tags": ["sessions"],
          "summary": "Get all sessions",
          "operationId": "getSessions",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/SessionOut"
                    },
                    "type": "array",
                    "title": "Response Getsessions"
                  }
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      },
      "/api/v1/session/{workout_id}": {
        "post": {
          "tags": ["sessions"],
          "summary": "Create an session",
          "operationId": "createSession",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Session"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/session/{session_id}": {
        "get": {
          "tags": ["sessions"],
          "summary": "Get a single session",
          "operationId": "getSession",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "session_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Session Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SessionOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["sessions"],
          "summary": "Update an session",
          "operationId": "updateSession",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "session_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Session Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SessionOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["sessions"],
          "summary": "Delete an session",
          "operationId": "deleteSession",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "session_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Session Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/session/active/sessions": {
        "get": {
          "tags": ["sessions"],
          "summary": "Get all active sessions",
          "operationId": "getActiveSessions",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/SessionOut"
                    },
                    "type": "array",
                    "title": "Response Getactivesessions"
                  }
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      },
      "/api/v1/session/workout/{workout_id}": {
        "get": {
          "tags": ["sessions"],
          "summary": "Get all sessions for a workout",
          "operationId": "getWorkoutSessions",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SessionWorkoutOut"
                    },
                    "title": "Response Getworkoutsessions"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Goal": {
          "properties": {
            "_id": {
              "anyOf": [
                {
                  "type": "string",
                  "example": "5eb7cf5a86d9755df3a6c593"
                },
                {
                  "type": "null"
                }
              ],
              "title": " Id",
              "description": "MongoDB document ObjectID"
            },
            "goal_id": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id",
              "unique": true
            },
            "category": {
              "$ref": "#/components/schemas/GoalType"
            },
            "goal": {
              "type": "integer",
              "title": "Goal"
            },
            "exercise": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exercise"
            },
            "owner": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Owner"
            }
          },
          "type": "object",
          "required": ["category", "goal", "owner"],
          "title": "Goal"
        },
        "GoalCreate": {
          "properties": {
            "category": {
              "$ref": "#/components/schemas/GoalType"
            },
            "goal": {
              "type": "integer",
              "title": "Goal"
            },
            "exercise": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exercise"
            }
          },
          "type": "object",
          "required": ["category", "goal", "exercise"],
          "title": "GoalCreate"
        },
        "GoalOut": {
          "properties": {
            "goal_id": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            },
            "category": {
              "$ref": "#/components/schemas/GoalType"
            },
            "goal": {
              "type": "integer",
              "title": "Goal"
            },
            "exercise": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exercise"
            },
            "owner": {
              "$ref": "#/components/schemas/UserOut"
            }
          },
          "type": "object",
          "required": ["goal_id", "category", "goal", "exercise", "owner"],
          "title": "GoalOut"
        },
        "GoalType": {
          "type": "string",
          "enum": ["weight", "pr", "completed"],
          "title": "GoalType"
        },
        "GoalUpdate": {
          "properties": {
            "category": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/GoalType"
                },
                {
                  "type": "null"
                }
              ]
            },
            "goal": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Goal"
            },
            "exercise": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exercise"
            }
          },
          "type": "object",
          "required": ["category", "goal", "exercise"],
          "title": "GoalUpdate"
        },
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
        "Session": {
          "properties": {
            "_id": {
              "anyOf": [
                {
                  "type": "string",
                  "example": "5eb7cf5a86d9755df3a6c593"
                },
                {
                  "type": "null"
                }
              ],
              "title": " Id",
              "description": "MongoDB document ObjectID"
            },
            "session_id": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id",
              "unique": true
            },
            "date": {
              "type": "string",
              "format": "date-time",
              "title": "Date"
            },
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "active": {
              "type": "boolean",
              "title": "Active",
              "default": false
            },
            "workout": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Workout"
            },
            "owner": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Owner"
            }
          },
          "type": "object",
          "required": ["workout", "owner"],
          "title": "Session"
        },
        "SessionCreate": {
          "properties": {
            "date": {
              "type": "string",
              "format": "date-time",
              "title": "Date"
            },
            "active": {
              "type": "boolean",
              "title": "Active"
            }
          },
          "type": "object",
          "required": ["date", "active"],
          "title": "SessionCreate"
        },
        "SessionOut": {
          "properties": {
            "session_id": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            },
            "date": {
              "type": "string",
              "format": "date-time",
              "title": "Date"
            },
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "active": {
              "type": "boolean",
              "title": "Active"
            },
            "workout": {
              "$ref": "#/components/schemas/Workout"
            }
          },
          "type": "object",
          "required": [
            "session_id",
            "date",
            "duration_mins",
            "active",
            "workout"
          ],
          "title": "SessionOut"
        },
        "SessionUpdate": {
          "properties": {
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "active": {
              "type": "boolean",
              "title": "Active"
            }
          },
          "type": "object",
          "required": ["duration_mins", "active"],
          "title": "SessionUpdate"
        },
        "SessionWorkoutOut": {
          "properties": {
            "session_id": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            },
            "date": {
              "type": "string",
              "format": "date-time",
              "title": "Date"
            },
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "active": {
              "type": "boolean",
              "title": "Active"
            }
          },
          "type": "object",
          "required": ["session_id", "date", "duration_mins", "active"],
          "title": "SessionWorkoutOut"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": ["loc", "msg", "type"],
          "title": "ValidationError"
        },
...
        "WorkoutOut": {
          "properties": {
            "workout_id": {
              "type": "string",
              "format": "uuid",
              "title": "Workout Id"
            },
            "title": {
              "type": "string",
              "title": "Title"
            },
            "created_date": {
              "type": "string",
              "format": "date-time",
              "title": "Created Date"
            },
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "day_of_week": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Day Of Week",
              "default": []
            }
          },
          "type": "object",
          "required": ["workout_id", "title", "created_date", "duration_mins"],
          "title": "WorkoutOut"
        },
        "WorkoutUpdate": {
          "properties": {
            "title": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50,
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Title"
            },
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "day_of_week": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Day Of Week",
              "default": []
            }
          },
          "type": "object",
          "required": ["title", "duration_mins"],
          "title": "WorkoutUpdate"
        }
      },
      "securitySchemes": {
        "JWT": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {},
              "tokenUrl": "/api/v1/auth/login"
            }
          }
        }
      }
    },
    "servers": [
      {
        "url": "https://shapeshift-backend.onrender.com"
      }
    ]
  }